% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{arrange_covars}
\alias{arrange_covars}
\title{Prepare covariates for \code{scones}}
\usage{
arrange_covars(gwas, covars)
}
\arguments{
\item{gwas}{A SnpMatrix object with the GWAS information.}

\item{covars}{A data frame with the covariates. It must contain a column 
'sample' containing the sample IDs, and an additional columns for each 
covariate.}
}
\value{
The covars data.frame, with the rows in the same order as gwas.
}
\description{
Prepares de covariates data.frame for the functions used in
\code{scones}, like \code{single_snp_association} or \code{score_folds}
.
}
\keyword{internal}
