% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/athal.R
\docType{data}
\name{athal}
\alias{athal}
\title{Description of the athal dataset.}
\format{A list with 3 items:
\describe{
  \item{gwas}{Genotype and phenotype information. Genotype was obtained by running \code{read.pedfile("genotype.ped", snps = "genotype.map")}. The phenotype was simulated using \code{simulate_phenotype(athal$gwas, athal$causal, h2=1, ncases=653, ncontrols=654, qualitative=T)}.}
  \item{net}{GM network of the genotyped SNPs in A. thaliana. Obtained from \code{get_GM_network(athal$gwas, 3702)}.}
  \item{causal}{Simulated causal SNPs generated by \code{simulate_causal_snps(athal$net, 60)}.}
  \item{cones}{SConES results with the best hyperparameters.}
}}
\description{
Simulated phenotypes over A. thaliana dataset from Horton et al. downloaded from \url{https://easygwas.ethz.ch/down/dataset/download/1/}.
}
\references{
Horton, M. W., Hancock, A. M., Huang, Y. S., Toomajian, C., Atwell, S., Auton, A., ... Bergelson, J. (2012). Genome-wide patterns of genetic variation in worldwide Arabidopsis thaliana accessions from the RegMap panel. Nature Genetics, 44(2), 212-216. \url{http://doi.org/10.1038/ng.1042}
}
