% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{calculateE}
\alias{calculateE}
\title{Calculate the environmental component of the phenotype}
\usage{
calculateE(G, h2)
}
\arguments{
\item{G}{The genetic component of the phenotype.}

\item{h2}{The heritability.}
}
\value{
A vector with the environmental component of each sample.
}
\description{
Calculates the environmental component of the phenotype using
the variance in the genetic component.
}
\keyword{internal}
