% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{calculateG}
\alias{calculateG}
\title{Calculate the genetic component of the phenotype}
\usage{
calculateG(u, X, model)
}
\arguments{
\item{u}{A vector with the effect size of each SNP.}

\item{X}{Genotypes.}

\item{model}{Genetic model to assume.}
}
\value{
A vector with the genetic component of each sample.
}
\description{
Calculates the genetic component of the phenotype from a genotype.
}
