% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_installed}
\alias{check_installed}
\title{Check package is installed}
\usage{
check_installed(pkgs, fn = "This function")
}
\arguments{
\item{pkgs}{Character vector with the names of the packages.}

\item{fn}{Function calling the check.}
}
\value{
The package is loaded into the namespace.
}
\description{
Checks if a package is installed, launches an error if it is
not.
}
\examples{
martini:::check_installed(c("martini"))
\dontrun{martini:::check_installed("martinid")}
}
\keyword{internal}
