% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_snps.R
\name{cluster_snps}
\alias{cluster_snps}
\title{Return groups of interconnected SNPs.}
\usage{
cluster_snps(map, net)
}
\arguments{
\item{map}{Results from \code{shake}.}

\item{net}{The same SNP network provided to \code{shake}.}
}
\value{
A list with the clusters of selected SNPs.
}
\description{
Find clusters composed by interconnected SNPs.
}
