% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{connect_biomart}
\alias{connect_biomart}
\title{Open a biomaRt connection}
\usage{
connect_biomart(organism)
}
\arguments{
\item{organism}{String containing the ensembl species name (e.g. hsapiens 
for human)}
}
\description{
Opens a biomaRt connection for the relevant species.
}
\keyword{internal}
