% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{encode_gwas}
\alias{encode_gwas}
\title{Change encoding of dataset}
\usage{
encode_gwas(X, encoding)
}
\arguments{
\item{X}{Genotype matrix with SNPs encoded as 0 for major homozygous, 1 for
heterozygous and 2 for minor homozygous.}

\item{encoding}{Genetic model assumed: additive, recessive, dominant or
codominant.}
}
\value{
A genotype matrix
}
\description{
Converts the encoding from additive to another one.
}
\examples{
X <- as(minigwas[["genotypes"]], "numeric")
martini:::encode_gwas(X, "recessive")
}
