% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{examplegwas}
\alias{examplegwas}
\title{Description of the examplegwas data.}
\format{A list with 3 items:
\describe{
  \item{gwas}{Genotype and phenotype information.}
  \item{net}{Simulated network.}
  \item{cones}{Result of runing \code{find_cones} with gwas and net.}
}}
\description{
Simulated genotype of 1800 SNPs over 6000 samples (3000 cases and 3000 controls).
}
