% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_snps.R
\name{find_snps}
\alias{find_snps}
\title{GWAS incorporating networks.}
\usage{
find_snps(gwas, net, ...)
}
\arguments{
\item{gwas}{A SnpMatrix object with the GWAS information.}

\item{net}{An igraph network that connects the SNPs.}

\item{...}{Extra arguments for shake.}
}
\value{
A copy of the SnpMatrix$map object, with the following additions:
\itemize{
\item{C: contains the univariate association score for every single SNP.}
\item{selected: logical vector indicating if the SNP was selected by shake or not.}
\item{cluster: integer with the number of the cluster the SNP belongs to.}
}
}
\description{
Finds the SNPs maximally associated with a phenotype while being connected in an underlying network (Azencott et al., 2013).
}
\references{
Azencott, C. A., Grimm, D., Sugiyama, M., Kawahara, Y., & Borgwardt, K. M. (2013). Efficient network-guided multi-locus association mapping with graph cuts. Bioinformatics, 29(13), 171–179. \url{https://doi.org/10.1093/bioinformatics/btt238}
}
