% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigmod.R
\name{get_A}
\alias{get_A}
\title{Compute adjacency matrix}
\usage{
get_A(gwas, net)
}
\arguments{
\item{gwas}{A SnpMatrix object with the GWAS information.}

\item{net}{An igraph network that connects the SNPs.}
}
\value{
An adjacency matrix.
}
\description{
Compute adjacency matrix
}
\keyword{internal}
