% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_adjacency}
\alias{get_adjacency}
\title{Compute Laplacian matrix}
\usage{
get_adjacency(gwas, net)
}
\arguments{
\item{gwas}{A SnpMatrix object with the GWAS information.}

\item{net}{An igraph network that connects the SNPs.}
}
\value{
A Laplacian matrix.
}
\description{
Compute Laplacian matrix
}
\keyword{internal}
