% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networks.R
\name{get_ensembl_genes}
\alias{get_ensembl_genes}
\title{Get genomic position of the genes.}
\usage{
get_ensembl_genes(gwas, organism = 9606)
}
\arguments{
\item{gwas}{A SnpMatrix object with the GWAS information.}
}
\value{
A data.frame with four columns, corresponding to the gene, the 
chromosome, the start and the end.
}
\description{
Get the genomic position of all the genes of an organism.
}
\keyword{internal}
