% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networks.R
\name{get_feature_network}
\alias{get_feature_network}
\title{Get feature membership network.}
\usage{
get_feature_network(gwas, snpMapping, col_features = c("snp", "feature"),
  feature_type = "feature")
}
\arguments{
\item{gwas}{A SnpMatrix object with the GWAS information.}

\item{snpMapping}{A data.frame informing how SNPs map to features It contains
at least two columns: SNP id and a feature it maps to. Each row corresponds 
to one gene-SNP mapping. Unless column names are specified using 
\code{col_features}, involved columns must be named \code{'snp'} and 
\code{'feature'}.}

\item{col_features}{Optional, length-2 character vector with the names of the 
two columns involving the SNP-gene mapping. The first element is the column 
of the SNP, and the second is the column of the gene.}

\item{feature_type}{Feature type.}
}
\value{
An igraph network where SNPs mapping to the same feature form 
cliques.
}
\description{
Creates a network of SNPs where each SNP is connected to all the
other SNPs mapped to the same feature.
}
