% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_GI_network.R
\name{get_GI_network}
\alias{get_GI_network}
\title{Get GI network.}
\usage{
get_GI_network(map, snp2gene, tab)
}
\arguments{
\item{map}{A map object with the SNP position information.}

\item{snp2gene}{A data frame with two columns: snp and gene. The first column contains the id of the SNP; the second, the gene which is mapped to.}

\item{tab}{A data frame describing protein-protein interactions in TAB format.}
}
\value{
An igraph network of the GI network of the SNPs.
}
\description{
Creates a network of SNPs where each SNP is connected as in the GM network and, in addition, to all the other SNPs pertaining to any interactor of the gene it is mapped to. Corresponds to the GI network described by Azencott et al.
}
\references{
Azencott, C. A., Grimm, D., Sugiyama, M., Kawahara, Y., & Borgwardt, K. M. (2013). Efficient network-guided multi-locus association mapping with graph cuts. Bioinformatics, 29(13), 171–179. \url{https://doi.org/10.1093/bioinformatics/btt238}
}
