% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_GS_network.R
\name{get_GS_network}
\alias{get_GS_network}
\title{Get GS network.}
\usage{
get_GS_network(map)
}
\arguments{
\item{map}{A map object with the SNP position information.}
}
\value{
An igraph network of the GS network of the SNPs.
}
\description{
Creates a network of SNPs where each SNP is connected to its adjacent SNPs in the genome sequence. Corresponds to the GS network described by Azencott et al.
}
\references{
Azencott, C. A., Grimm, D., Sugiyama, M., Kawahara, Y., & Borgwardt, K. M. (2013). Efficient network-guided multi-locus association mapping with graph cuts. Bioinformatics, 29(13), 171–179. \url{https://doi.org/10.1093/bioinformatics/btt238}
}
