% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networks.R
\name{get_gxg}
\alias{get_gxg}
\title{Get gene interactions}
\usage{
get_gxg(db, organism, flush)
}
\arguments{
\item{db}{String containing the database to obtain the gene-gene interactions
from. Possible values: 'biogrid', 'string'.}

\item{organism}{Organism: 9606 represents human, etc.}

\item{flush}{Remove cached results? Boolean value.}
}
\description{
Wrapper for the different functions to get gene-gene 
interactions. Supports cached results.
}
\keyword{internal}
