% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networks.R
\name{get_gxg_biogrid}
\alias{get_gxg_biogrid}
\title{Get BioGRID protein-protein interactions.}
\usage{
get_gxg_biogrid(organism = 9606)
}
\arguments{
\item{organism}{Organism: human represents human, arabidopsis for Arabidopsis
thaliana, etc.}
}
\value{
A data.frame with two columns with pairs of interacting proteins.
}
\description{
Get all protein-protein interactions for an organism from
BioGRID.
}
\examples{
# download dog interactions
\dontrun{martini:::get_gxg_biogrid(9615)}
}
\keyword{internal}
