% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networks.R
\name{get_gxg_string}
\alias{get_gxg_string}
\title{Get STRING protein-protein interactions.}
\usage{
get_gxg_string(organism = 9606)
}
\arguments{
\item{organism}{Organism: 9606 represents human, etc.}
}
\value{
A data.frame with two columns with pairs of interacting proteins.
}
\description{
Get all protein-protein interactions for an organism from
STRING. It uses a score cut-off of 400.
}
\examples{
# download frog interactions
\dontrun{martini:::get_gxg_string(8364)}
}
\keyword{internal}
