% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ld.R
\name{get_ld}
\alias{get_ld}
\title{Retrieve LD information from Ensembl}
\usage{
get_ld(net)
}
\arguments{
\item{net}{A SNP network.}
}
\value{
An dataframe with a column key, with SNP ids separated by an underscore, and an r2 column, 
containing the Pearson correlation.
}
\description{
Retrieve LD information from Ensembl in the format required by ldweight_edges.
}
