% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ld.R
\name{get_ld_from_gwas}
\alias{get_ld_from_gwas}
\title{Retrieve LD information from Ensembl}
\usage{
get_ld_from_gwas(gwas, organism, db)
}
\arguments{
\item{gwas}{A GWAS experiment, in snpMatrix form.}
}
\value{
An dataframe with a column key, with SNP ids separated by an underscore, and an r2 column, 
containing the Pearson correlation.
}
\description{
Retrieve LD information from Ensembl in the format required by ldweight_edges.
}
