% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{get_ppi}
\alias{get_ppi}
\title{Get protein-protein interactions.}
\usage{
get_ppi(organism = 9606)
}
\arguments{
\item{organism}{Organism: human represents human, arabidopsis for Arabidopsis
thaliana, etc.}
}
\value{
A data.frame with two columns with pairs of interacting proteins.
}
\description{
Get all protein-protein interactions for an organism from
BioGRID.
}
\examples{
# download dog interactions
martini:::get_ppi(9615)
}
\keyword{internal}
