% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_shake_settings.R
\name{get_shake_settings}
\alias{get_shake_settings}
\title{Get shake settings.}
\usage{
get_shake_settings(...)
}
\arguments{
\item{...}{Any \code{shake} option.}
}
\value{
A list of \code{shake} settings.
}
\description{
Creates a list composed by all \code{shake} settings, with the values provided by the user, or the default ones if none is 
provided.
}
