% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evo.R
\name{get_snp_modules}
\alias{get_snp_modules}
\title{Return groups of interconnected SNPs.}
\usage{
get_snp_modules(cones, net)
}
\arguments{
\item{cones}{Results from \code{evo}.}

\item{net}{The same SNP network provided to \code{evo}.}
}
\value{
A list with the modules of selected SNPs.
}
\description{
Find modules composed by interconnected SNPs.
}
\examples{
gi <- get_GI_network(minigwas, snpMapping = minisnpMapping, ppi = minippi)
cones <- search_cones(minigwas, gi)
get_snp_modules(cones, gi)
}
