% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{group_snps}
\alias{group_snps}
\title{Groups nearby SNPs}
\usage{
group_snps(bed, chr_col, pos_col, threshold)
}
\arguments{
\item{bed}{data.frame containing at least two properties (chromosome and 
position) of a set of SNPs.}

\item{chr_col}{Name of the column containing the SNP chromosome.}

\item{pos_col}{Name of the column containing the SNP position.}

\item{threshold}{Maximum distance to group two SNPs group.}
}
\value{
A data.frame in bed format, with the same dimensions as the original,
but with the groups.
}
\description{
Groups SNPs closer than a specifiec threshold of distance.
}
