% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ld.R
\name{ld_prune}
\alias{ld_prune}
\title{LD-remove redundant SNPs}
\usage{
ld_prune(gwas, ld, cutoff = 0.8)
}
\arguments{
\item{gwas}{A GWAS experiment.}

\item{ld}{A matrix of # SNPs x # SNPs representing the LD between pairs of 
SNPs. For this function to work, only the LD with the adjacent SNPs is 
required.}

\item{cutoff}{Minimum LD to be considered part of the same LD-block.}
}
\value{
A copy of the GWAS experiment without redundant SNPs.
}
\description{
Removes SNPs belonging to the same LD-block, leaving only one. 
An LD-block is defined as sets of SNPs that have a LD measure with the 
adjacent SNP higher than the cut-off.
}
\examples{
ld <- snpStats::ld(minigwas$genotypes, depth = 2, stats = "R.squared")
ld_prune(minigwas, ld)
}
