% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ld.R
\name{ldweight_edges}
\alias{ldweight_edges}
\title{Include LD information in the network.}
\usage{
ldweight_edges(net, ld = get_ld(net))
}
\arguments{
\item{net}{A SNP network.}
}
\value{
An SNP network where the edges weight 1 - LD, measured as Pearson correlation.
}
\description{
Include LD information in the SNP network. Only work for human SNPs.
}
