% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{maxflow}
\alias{maxflow}
\title{Maxflow algorithm}
\usage{
maxflow(A, W)
}
\arguments{
\item{A}{A matrix with two columns. The first column contains the edges to
the source; the second, the edges to the sink.}

\item{W}{A sparse matrix with the connectivity.}
}
\value{
A list with vector indicating if the feature was selected and the 
objective score.
}
\description{
Run the maxflow algorithm.
}
