% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epistasis.R
\name{measure_epistasis}
\alias{measure_epistasis}
\title{Measure epistasis between SNPs}
\usage{
measure_epistasis(gwas, net)
}
\arguments{
\item{gwas}{A GWAS experiment.}

\item{net}{A network of SNPs.}
}
\value{
A dataframe with the chi-squared score for each pair of tested SNPs.
}
\description{
Measure epistasis between SNPs along the edges of net.
}
