% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{mget_gxg_biogrid}
\alias{mget_gxg_biogrid}
\title{Memoised version of get_gxg_biogrid}
\usage{
mget_gxg_biogrid(organism = 9606)
}
\arguments{
\item{organism}{Tax ID of the studied organism. The default is 9606 (human).}
}
\value{
A data.frame with two columns with pairs of interacting proteins.
}
\description{
Get all protein-protein interactions for an organism from
BioGRID.
}
\examples{
# download dog interactions
\dontrun{martini:::get_gxg_biogrid(9615)}
}
\keyword{internal}
