% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{mget_gxg_string}
\alias{mget_gxg_string}
\title{Memoised version of get_gxg_stringdb}
\usage{
mget_gxg_string(organism = 9606)
}
\arguments{
\item{organism}{Tax ID of the studied organism. The default is 9606 (human).}
}
\value{
A data.frame with two columns with pairs of interacting proteins.
}
\description{
Get all protein-protein interactions for an organism from
STRING. It uses a score cut-off of 400.
}
\examples{
# download frog interactions
\dontrun{martini:::get_gxg_string(8364)}
}
\keyword{internal}
