% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scones.R
\name{mincut.cv}
\alias{mincut.cv}
\title{Run the cross-validated min-cut algorithm}
\usage{
mincut.cv(gwas, net, net_matrix, covars, opts)
}
\arguments{
\item{gwas}{A SnpMatrix object with the GWAS information.}

\item{net}{An igraph network that connects the SNPs.}
}
\description{
Run the cross-validated min-cut algorithm
}
\keyword{internal}
