% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scones.R
\name{mincut}
\alias{mincut}
\title{Run min-cut algorithm}
\usage{
mincut(gwas, net, covars, eta, lambda, score, sigmod, family, link)
}
\value{
A copy of the \code{SnpMatrix$map} \code{data.frame}, with the 
following additions:
\itemize{
\item{c: contains the univariate association score for every single SNP.}
\item{selected: logical vector indicating if the SNP was selected by SConES 
or not.}
\item{module: integer with the number of the module the SNP belongs to.}
}
}
\description{
Run min-cut algorithm
}
\keyword{internal}
