% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{minigwas}
\alias{minigwas}
\title{Description of the minigwas dataset.}
\format{A list with 3 items:
\describe{
  \item{genotypes}{Genotype and phenotype information.}
  \item{fam}{Simulated network.}
  \item{map}{Result of runing \code{find_cones} with gwas and net.}
}}
\description{
Small GWAS example.
}
\examples{
data(minigwas)

# access different elements
minigwas$genotypes
minigwas$map
minigwas$fam
}
