% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{minippi}
\alias{minippi}
\title{PPIs for the minigwas dataset.}
\description{
data.frame describing pairs of proteins that interact for 
minigwas.
}
\examples{
data(minippi)

head(minippi)
}
