% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{minisnpMapping}
\alias{minisnpMapping}
\title{Genes for the minigwas dataset.}
\description{
Data frame that maps SNPs from minigwas to their gene.
}
\examples{
data(minisnpMapping)

head(minisnpMapping)
}
