% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scones.R
\name{parse_scones_settings}
\alias{parse_scones_settings}
\title{Parse search_cones settings.}
\usage{
parse_scones_settings(associationScore = "chi2",
  modelScore = "consistency", etas = numeric(), lambdas = numeric(),
  debug = FALSE, c = numeric())
}
\arguments{
\item{associationScore}{Association score to measure association between 
genotype and phenotype. Possible values: chi2 (default), glm.}

\item{modelScore}{Model selection criterion Possible values: consistency, 
bic (default), aic, aicc, mbic.}

\item{etas}{Numeric vector with the etas to explore in the grid search. If 
ommited, it's automatically created based on the association
scores.}

\item{lambdas}{Numeric vector with the lambdas to explore in the grid search.
If ommited, it's automatically created based on the association scores.}

\item{debug}{Display additional information. Possible values: TRUE, FALSE
(default).}

\item{c}{Numeric vector with the association scores of the SNPs. Specify it 
to automatically an appropriate range of etas and lambas.}
}
\value{
A list of \code{evo} settings.
}
\description{
Creates a list composed by all \code{evo} settings, with the 
values provided by the user, or the default ones if none is provided.
}
\examples{
martini:::parse_scones_settings(etas = c(1,2,3), lambdas = c(4,5,6))
martini:::parse_scones_settings(c = c(1,10,100), associationScore = "skat")
}
\keyword{internal}
