% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigmod.R
\name{parse_sigmod_settings}
\alias{parse_sigmod_settings}
\title{Parse \code{sigmod.cv} settings}
\usage{
parse_sigmod_settings(gwas, covars = data.frame(), c = as.numeric(), ...)
}
\arguments{
\item{gwas}{A SnpMatrix object with the GWAS information.}

\item{covars}{A data frame with the covariates. It must contain a column 
'sample' containing the sample IDs, and an additional columns for each 
covariate.}

\item{c}{Numeric vector with the association scores of the SNPs. Specify it 
to automatically an appropriate range of etas and lambas.}

\item{score}{Association score to measure association between genotype and 
phenotype. Possible values: chi2 (default), glm.}

\item{criterion}{String with the function to measure the quality of a split. 
Possible values: consistency (default), bic, aic, aicc.}

\item{etas}{Numeric vector with the etas to explore in the grid search. If 
ommited, it's automatically created based on the association
scores.}

\item{lambdas}{Numeric vector with the lambdas to explore in the grid search.
If ommited, it's automatically created based on the association scores.}
}
\value{
A list of \code{sigmod.cv} settings.
}
\description{
Creates a list composed by all \code{sigmod.cv} settings, with 
the values provided by the user, or the default ones if none is provided.
}
\examples{
martini:::parse_sigmod_settings(minigwas, etas = c(1,2,3), lambdas = c(4,5,6))
martini:::parse_sigmod_settings(minigwas, c = c(1,10,100), score = "glm")
}
\keyword{internal}
