% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{permute_snpMatrix}
\alias{permute_snpMatrix}
\title{Permute samples}
\usage{
permute_snpMatrix(gwas)
}
\arguments{
\item{gwas}{A SnpMatrix object with the GWAS information.}
}
\description{
Compute a permutation of the samples of a snpMatrix object. 
Useful to make sure that the folds are not stratified by phenotype.
}
\keyword{internal}
