% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_ideogram}
\alias{plot_ideogram}
\title{Ideogram of SConES results.}
\usage{
plot_ideogram(cones, genome = "hg19")
}
\arguments{
\item{cones}{Output from \code{SConES}.}

\item{genome}{Abbreviations of the genome to use: hg19 for human (default), mm10 for mouse, etc. Argument to be passed to 
\code{\link{circos.initializeWithIdeogram}} \code{species}.}
}
\value{
A circular ideogram, including the manhattan plot, and the interactions between the selected SNPs.
}
\description{
Create a circular ideogram of the \code{SConES} results using the circlize package (Gu et al., 2014).
}
\references{
Gu, Z., Gu, L., Eils, R., Schlesner, M., & Brors, B. (2014). circlize Implements and enhances circular visualization in R. 
Bioinformatics (Oxford, England), 30(19), 2811-2. \url{https://doi.org/10.1093/bioinformatics/btu393}
}
