% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_snp_cluster.R
\name{plot_snp_cluster}
\alias{plot_snp_cluster}
\title{Ideogram of a SNP cluster.}
\usage{
plot_snp_cluster(cones, k, genome = "hg19")
}
\arguments{
\item{cones}{Results from \code{shake}.}

\item{k}{Id of the cluster to plot.}

\item{genome}{Abbreviations of the genome to use: hg19 for human (default),  mm10 for mouse, etc. Argument to be passed to \code{\link{IdeogramTrack}}.}
}
\value{
An ideogram per chromosome showing the selected SNPs and the genes in the region.
}
\description{
Create a ideogram of a SNP cluster from \code{shake} results using the Gviz package (Hahne and Ivanek, 2016).
}
\references{
Hahne F. and Ivanek R. (2016). "Statistical Genomics: Methods and Protocols." In Mathe E and Davis S (eds.), chapter Visualizing Genomic Data Using Gviz and Bioconductor, pp. 335-351. Springer New York, New York, NY. ISBN 978-1-4939-3578-9, doi: 10.1007/978-1-4939-3578-9_16, \url{http://dx.doi.org/10.1007/978-1-4939-3578-9_16}.
}
