% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_snp_module}
\alias{plot_snp_module}
\title{Ideogram of a SNP module.}
\usage{
plot_snp_module(cones, k, genome = "hg19")
}
\arguments{
\item{cones}{Results from \code{SConES}.}

\item{k}{Id of the module to plot.}

\item{genome}{Abbreviations of the genome to use: hg19 for human (default),  mm10 for mouse, etc. Argument to be passed to 
\code{\link[Gviz]{IdeogramTrack}}.}
}
\value{
An ideogram per chromosome showing the selected SNPs and the genes in the region.
}
\description{
Create a ideogram of a SNP module from \code{SConES} results using the Gviz package (Hahne and Ivanek, 2016).
}
\references{
Hahne F. and Ivanek R. (2016). "Statistical Genomics: Methods and Protocols." In Mathe E and Davis S (eds.), chapter 
Visualizing Genomic Data Using Gviz and Bioconductor, pp. 335-351. Springer New York, New York, NY. ISBN 978-1-4939-3578-9, doi: 
10.1007/978-1-4939-3578-9_16, \url{http://dx.doi.org/10.1007/978-1-4939-3578-9_16}.
}
