% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{run_scones}
\alias{run_scones}
\title{Run shake.}
\usage{
run_scones(c, eta, lambda, W)
}
\arguments{
\item{c}{A vector with the association of each SNP with the phenotype.}

\item{eta}{A numeric with the value of the eta parameter.}

\item{lambda}{A numeric with the value of the eta parameter.}

\item{W}{A sparse matrix with the connectivity.}
}
\value{
A list with vector indicating if the feature was selected and the 
objective score.
}
\description{
Run scones.
}
