% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{run_shake}
\alias{run_shake}
\title{Run shake.}
\usage{
run_shake(X, Y, network, userSettings)
}
\arguments{
\item{X}{A matrix with the genotypes.}

\item{Y}{A vector with the phenptypes.}

\item{network}{A sparse matrix containing the adjacency matrix of the network.}

\item{userSettings}{A named list with the settings.}
}
\value{
An object with the shake results.
}
\description{
Run shake.
}
