% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sanitize_map}
\alias{sanitize_map}
\title{Check map}
\usage{
sanitize_map(gwas)
}
\arguments{
\item{gwas}{A SnpMatrix object with the GWAS information.}
}
\description{
Check that map is a proper data.frame.
}
\keyword{internal}
