% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sanitize_snpMapping}
\alias{sanitize_snpMapping}
\title{Check snpMapping}
\usage{
sanitize_snpMapping(snpMapping, col_genes)
}
\arguments{
\item{snpMapping}{data.frame containing the correspondence between SNPs and
genes.}

\item{col_genes}{Length 2 character vector containing the colnames containing
the SNP and the gene ids, respectively.}
}
\description{
Check that snpMapping is a proper data.frame.
}
\keyword{internal}
