% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scones.R
\name{score_fold}
\alias{score_fold}
\title{Score the solutions of a k-fold}
\usage{
score_fold(folds, modelScore)
}
\arguments{
\item{folds}{k \times d matrix, where k is the number of folds, and d the 
number of SNPs.}

\item{modelScore}{String with the method to use to score the folds.}
}
\description{
Take the k-solutions for a combination of hyperparameters, and 
assign a score to it (the larger, the better).
}
\keyword{internal}
