% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shake_test.R
\name{shake_test}
\alias{shake_test}
\title{Calculate an empirical p-value for shake results.}
\usage{
shake_test(map, net, n = 1e+05)
}
\arguments{
\item{map}{Results from \code{shake}.}

\item{net}{The same SNP network provided to \code{shake}.}

\item{n}{Integer with the name of permutations.}
}
\value{
An empirical p-value for each of the SNP clusters.
}
\description{
Do a permutation-based test to assess the statistical significance of each of the clusters obtained through shake.
}
