% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_phenotype.R
\name{simulate_phenotype}
\alias{simulate_phenotype}
\title{Simulate phenotype}
\usage{
simulate_phenotype(X, snps, h2, model = "additive",
  effectSize = rnorm(sum(snps)), qualitative = FALSE, ncases, ncontrols)
}
\arguments{
\item{X}{A numeric matrix, rows are samples, columns are SNPs.}

\item{snps}{Boolean vector identifying the causal SNPs. Its length must be equal to the number of columns of X.}

\item{h2}{Heritability of the phenotype (between 0 and 1).}

\item{model}{String specifying the genetic model under the phenotype. Accepted values: "additive".}

\item{effectSize}{Numeric vector with the same lenght as the number of causal SNPs. It indicates the effect size of each of the SNPs; if absent, they are sampled fron a normal distribution.}

\item{qualitative}{Bool indicating if the phenotype is qualitative or not (quantitative).}

\item{ncases}{Integer specifying the number of cases to simulate in a qualitative phenotype. Required if qualitative = TRUE.}

\item{ncontrols}{Integer specifying the number of controls to simulate in a qualitative phenotype. Required if qualitative = TRUE.}
}
\value{
A vector with the simulated phenotype for each sample.
}
\description{
Simulates a phenotype from a GWAS experiment and a specified set of causal SNPs.
}
\references{
Inspired from GCTA simulation tool: \url{http://cnsgenomics.com/software/gcta/Simu.html}.
}
