% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networks.R
\name{snp2feature}
\alias{snp2feature}
\title{Map SNPs to genomic features.}
\usage{
snp2feature(gwas, features, flank = 0)
}
\arguments{
\item{gwas}{A SnpMatrix object with the GWAS information.}

\item{flank}{A number with the flanking regions around features to be 
considered part of the feature i.e. SNPs mapped to them will be considered 
mapped to the feature.}
}
\value{
A data.frame with two columns: one for the SNP and another for the
feature it has been mapped to.
}
\description{
Maps SNPs from a GWAS experiment to genomic features
}
\keyword{internal}
