% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networks.R
\name{snp2gene}
\alias{snp2gene}
\title{Map SNPs to genes.}
\usage{
snp2gene(gwas, organism = 9606, flank = 0)
}
\arguments{
\item{gwas}{A SnpMatrix object with the GWAS information.}

\item{organism}{Organism: hsapiens represents human, athaliana for
Arabidopsis thaliana, etc.}

\item{flank}{A number with the flanking regions around genes to be considered
part of the gene i.e. SNPs mapped to them will be considered mapped to the
gene.}
}
\value{
A dataframe with two columns: one for the SNP and another for the
gene it has been mapped to.
}
\description{
Maps SNPs from a GWAS experiment to genes.
}
\examples{
\dontrun{get_ppi(gwas, 9606, 50000)}
}
