% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{subnet}
\alias{subnet}
\title{Subgraph of vertices with an attribute}
\usage{
subnet(net, attr, values)
}
\arguments{
\item{net}{An igraph network.}

\item{attr}{An attribute of the vertices.}

\item{values}{Possible values of \code{attr}}
}
\value{
A subgraph containing only the vertices with attribute equal to any of the values in \code{values}.
}
\description{
Takes a map file and:
}
