% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{subset_snpMatrix}
\alias{subset_snpMatrix}
\title{Subsample snpMatrix}
\usage{
subset_snpMatrix(gwas, samples)
}
\arguments{
\item{gwas}{A SnpMatrix object with the GWAS information.}

\item{samples}{Vector (logical or numeric) containing the samples to select.}
}
\description{
Compute a permutation of the samples of a snpMatrix object. 
Useful to make sure that the folds are not stratified by phenotype.
}
\keyword{internal}
