% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_networks.R
\name{weight_edges_ld}
\alias{weight_edges_ld}
\title{Get GS network.}
\usage{
weight_edges_ld(net, organism = 9606)
}
\arguments{
\item{net}{A SNP network.}
}
\value{
An SNP network where the edges weight 1 - LD, measured as Pearson correlation.
}
\description{
Creates a network of SNPs where each SNP is connected to its adjacent SNPs in the genome sequence. Corresponds to the GS 
network described by Azencott et al.
}
