% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epistasis.R
\name{weight_epistasis_network}
\alias{weight_epistasis_network}
\title{Incorporate epistasis into the network}
\usage{
weight_epistasis_network(net, scores)
}
\arguments{
\item{net}{A network of SNPs.}

\item{scores}{A dataframe with pairs of SNPs and the measure of the epistatic
effect.}
}
\value{
A network where the weight of the edges reflect the strength of the 
epistatic interaction.
}
\description{
Incorporate epistatic interactions in the network.
}
