% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nongwas.R
\name{wrap_Xy}
\alias{wrap_Xy}
\title{Make pseudo SnpMatrix object}
\usage{
wrap_Xy(X, y, featnames, net)
}
\arguments{
\item{X}{n x d design matrix}

\item{y}{Vector of length n with the outcomes}

\item{featnames}{Vector of length d with the feature names}

\item{net}{An igraph network that connects the SNPs.}
}
\description{
Wrap design matrix and outcome vector into a pseudo SnpMatrix
object.
}
