% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{calculateG}
\alias{calculateG}
\title{Calculate the genetic component of the phenotype}
\usage{
calculateG(effectSize, X, model)
}
\arguments{
\item{effectSize}{A vector with the effect size of each SNP.}

\item{X}{Genotypes in a numeric matrix, where each row is a sample and each
column a SNP.}

\item{model}{Genetic model to assume.}
}
\value{
A vector with the genetic component of each sample.
}
\description{
Calculates the genetic component of the phenotype from a
genotype.
}
