% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_installed}
\alias{check_installed}
\title{Check package is installed}
\usage{
check_installed(pkg, fn = "this function")
}
\arguments{
\item{pkg}{Name of the package.}

\item{fn}{Function calling the check.}
}
\description{
Checks if a package is installed, launches an error if it is not.
}
