% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evo.R
\name{cluster_snps}
\alias{cluster_snps}
\title{Return groups of interconnected SNPs.}
\usage{
cluster_snps(cones, net)
}
\arguments{
\item{cones}{Results from \code{evo}.}

\item{net}{The same SNP network provided to \code{evo}.}
}
\value{
A list with the clusters of selected SNPs.
}
\description{
Find clusters composed by interconnected SNPs.
}
