% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_evo_settings.R
\name{get_evo_settings}
\alias{get_evo_settings}
\title{Get shake settings.}
\usage{
get_evo_settings(...)
}
\arguments{
\item{associationScore}{Association score to measure association between genotype and phenotype. Possible values: skat (default), chi2, trend.}

\item{modelScore}{Model selection criterion Possible values: consistency, bic (default), aic, aicc, mbic.}

\item{encoding}{SNP encoding. Possible values: additive (default), resessive, dominant, codominant.}

\item{debug}{Display additional information. Possible values: TRUE, FALSE (default).}
}
\value{
A list of \code{shake} settings.
}
\description{
Creates a list composed by all \code{shake} settings, with the values provided by the user, or the default ones if none is provided.
}
