% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legacy.R
\name{get_evo_settings}
\alias{get_evo_settings}
\title{Get evo settings.}
\usage{
get_evo_settings(
  associationScore = "chi2",
  modelScore = "bic",
  etas = numeric(),
  lambdas = numeric(),
  debug = FALSE
)
}
\arguments{
\item{associationScore}{Association score to measure association between 
genotype and phenotype. Possible values: chi2 (default), glm.}

\item{modelScore}{Model selection criterion Possible values: consistency, 
bic (default), aic, aicc, mbic.}

\item{etas}{Numeric vector with the etas to explore in the grid search. If 
ommited, it's automatically created based on the association
scores.}

\item{lambdas}{Numeric vector with the lambdas to explore in the grid search.
If ommited, it's automatically created based on the association scores.}

\item{debug}{Display additional information. Possible values: TRUE, FALSE
(default).}
}
\value{
A list of \code{evo} settings.
}
\description{
Creates a list composed by all \code{evo} settings, with the 
values provided by the user, or the default ones if none is provided.
}
\examples{
martini:::get_evo_settings()
martini:::get_evo_settings(associationScore = "skat")
}
\keyword{internal}
