% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_GM_network.R
\name{get_GM_network}
\alias{get_GM_network}
\title{Get GM network.}
\usage{
get_GM_network(map, snp2gene)
}
\arguments{
\item{map}{A map object with the SNP position information.}

\item{snp2gene}{A data frame with two columns: snp and gene. The first column contains the id of the SNP; the second, the gene which is mapped to.}
}
\value{
An igraph network of the GM network of the SNPs.
}
\description{
Creates a network of SNPs where each SNP is connected as in the GS network and, in addition, to all the other SNPs pertaining to the same gene. Corresponds to the GM network described by Azencott et al.
}
\references{
Azencott, C. A., Grimm, D., Sugiyama, M., Kawahara, Y., & Borgwardt, K. M. (2013). Efficient network-guided multi-locus association mapping with graph cuts. Bioinformatics, 29(13), 171–179. \url{https://doi.org/10.1093/bioinformatics/btt238}
}
