% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ld.R
\name{get_ld}
\alias{get_ld}
\title{Calculate LD in the datasets}
\usage{
get_ld(gwas, window = 50000)
}
\arguments{
\item{gwas}{A GWAS experiment, in snpMatrix form.}

\item{window}{Window size.}
}
\value{
An dataframe with a column key, with SNP ids separated by a "-"
character, and an r2 column, containing the Pearson correlation.
}
\description{
Calculate LD as pairwise correlations in the GWAS dataset.
Creates a sliding window for each SNP and calculates all the pairwise 
correlations inside that window.
}
