% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scones.R
\name{get_snp_modules}
\alias{get_snp_modules}
\title{Return groups of interconnected SNPs.}
\usage{
get_snp_modules(cones, net)
}
\arguments{
\item{cones}{Results from \code{scones.cv}.}

\item{net}{An igraph network that connects the SNPs.}
}
\value{
A list with the modules of selected SNPs.
}
\description{
Find modules composed by interconnected SNPs.
}
\examples{
gi <- get_GI_network(minigwas, snpMapping = minisnpMapping, ppi = minippi)
cones <- scones.cv(minigwas, gi)
martini:::get_snp_modules(cones, gi)
}
\keyword{internal}
