% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_coherent}
\alias{is_coherent}
\title{Check inner coherence of GWAS dataset}
\usage{
is_coherent(gwas)
}
\arguments{
\item{gwas}{A GWAS experiment.}
}
\value{
TRUE if the GWAS dataset is coherent. Else, raises an error.
}
\description{
Checks that the different data structures have the SNPs in the 
same order.
}
\examples{
martini:::is_coherent(minigwas)
}
\keyword{internal}
