% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ld.R
\name{ldweight_edges}
\alias{ldweight_edges}
\title{Include LD information in the network.}
\usage{
ldweight_edges(net, ld, method = "inverse")
}
\arguments{
\item{net}{A SNP network.}

\item{ld}{Data frame with linkage disequilibrium, like \code{snpStats::ld} 
output.}

\item{method}{How to incorporate LD values into the network.}
}
\value{
An SNP network where the edges weight 1 - LD, measured as Pearson
correlation.
}
\description{
Include LD information in the SNP network. Only work for human
SNPs.
}
\examples{
ld <- snpStats::ld(minigwas$genotypes, depth = 2, stats = "R.squared")
# don't weight edges for which LD cannot be calculated
ld[is.na(ld)] <- 0
gi <- get_GI_network(minigwas, snpMapping = minisnpMapping, ppi = minippi)
ldGi <- ldweight_edges(gi, ld)
}
