% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{maxflow}
\alias{maxflow}
\title{Maxflow algorithm}
\usage{
maxflow(A, As, At)
}
\arguments{
\item{A}{A sparse matrix with the connectivity.}

\item{As}{A vector containing the edges to the source.}

\item{At}{A vector containing the edges to the sink.}
}
\value{
A list with vector indicating if the feature was selected and the 
objective score.
}
\description{
Run the maxflow algorithm.
}
