% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scones.R
\name{mincut.cv}
\alias{mincut.cv}
\title{Run the cross-validated min-cut algorithm}
\usage{
mincut.cv(gwas, net, covars, lambdas, etas, criterion, score, sigmod)
}
\arguments{
\item{gwas}{A SnpMatrix object with the GWAS information.}

\item{net}{An igraph network that connects the SNPs.}

\item{covars}{A data frame with the covariates. It must contain a column 
'sample' containing the sample IDs, and an additional columns for each 
covariate.}
}
\description{
Run the cross-validated min-cut algorithm
}
\keyword{internal}
