% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{organism_id2name}
\alias{organism_id2name}
\title{Tax id to ensembl species name}
\usage{
organism_id2name(id)
}
\arguments{
\item{organism}{Tax ID of the studied organism. The default is 9606 (human).}
}
\description{
Converts taxid to ensembl species name e.g. human databases are 
hsapiens_*
}
\keyword{internal}
