% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scones.R
\name{parse_scones_settings}
\alias{parse_scones_settings}
\title{Parse \code{scones.cv} settings.}
\usage{
parse_scones_settings(
  c = numeric(),
  score = "chi2",
  criterion = "consistency",
  etas = numeric(),
  lambdas = numeric()
)
}
\arguments{
\item{c}{Numeric vector with the association scores of the SNPs. Specify it 
to automatically an appropriate range of etas and lambas.}

\item{score}{Association score to measure association between 
genotype and phenotype. Possible values: chi2 (default), glm.}

\item{criterion}{String with the function to measure the quality of a split. 
Possible values: consistency (default), bic, aic, aicc.}

\item{etas}{Numeric vector with the etas to explore in the grid search. If 
ommited, it's automatically created based on the association
scores.}

\item{lambdas}{Numeric vector with the lambdas to explore in the grid search.
If ommited, it's automatically created based on the association scores.}
}
\value{
A list of \code{evo} settings.
}
\description{
Creates a list composed by all \code{scones.cv} settings, with 
the values provided by the user, or the default ones if none is provided.
}
\examples{
martini:::parse_scones_settings(etas = c(1,2,3), lambdas = c(4,5,6))
martini:::parse_scones_settings(c = c(1,10,100), score = "glm")
}
\keyword{internal}
