% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ideogram.R
\name{plot_ideogram}
\alias{plot_ideogram}
\title{Ideogram of shake results.}
\usage{
plot_ideogram(map, net)
}
\arguments{
\item{map}{Output from \code{shake}.}

\item{net}{The SNP network provided to \code{shake}.}
}
\value{
A circular ideogram, including the manhattan plot, and the interactions between the selected SNPs.
}
\description{
Create a circular ideogram of the \code{shake} results using the circlize package (Gu et al., 2014).
}
\references{
Gu, Z., Gu, L., Eils, R., Schlesner, M., & Brors, B. (2014). circlize Implements and enhances circular visualization in R. Bioinformatics (Oxford, England), 30(19), 2811-2. \url{https://doi.org/10.1093/bioinformatics/btu393}
}
