% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scones.R
\name{scones.cv}
\alias{scones.cv}
\title{Find connected explanatory SNPs.}
\usage{
scones.cv(gwas, net, covars = data.frame(), ...)
}
\arguments{
\item{gwas}{A SnpMatrix object with the GWAS information.}

\item{net}{An igraph network that connects the SNPs.}

\item{covars}{A data frame with the covariates. It must contain a column 
'sample' containing the sample IDs, and an additional columns for each 
covariate.}

\item{...}{Extra arguments for \code{\link{parse_scones_settings}}.}
}
\value{
A copy of the \code{SnpMatrix$map} \code{data.frame}, with the 
following additions:
\itemize{
\item{c: contains the univariate association score for every single SNP.}
\item{selected: logical vector indicating if the SNP was selected by SConES 
or not.}
\item{module: integer with the number of the module the SNP belongs to.}
}
}
\description{
Finds the SNPs maximally associated with a phenotype while being
connected in an underlying network (Azencott et al., 2013). Select the 
hyperparameters by cross-validation.
}
\examples{
gi <- get_GI_network(minigwas, snpMapping = minisnpMapping, ppi = minippi)
scones.cv(minigwas, gi)
scones.cv(minigwas, gi, score = "glm")
}
\references{
Azencott, C. A., Grimm, D., Sugiyama, M., Kawahara, Y., & 
Borgwardt, K. M. (2013). Efficient network-guided multi-locus 
association mapping with graph cuts. Bioinformatics, 29(13), 171-179. 
\url{https://doi.org/10.1093/bioinformatics/btt238}
}
