% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scones.R
\name{score_fold}
\alias{score_fold}
\title{Score the solutions of a k-fold}
\usage{
score_fold(folds, criterion, K, gwas, covars)
}
\arguments{
\item{folds}{k-times-d matrix, where k is the number of folds, and d the 
number of SNPs.}

\item{criterion}{String with the method to use to score the folds.}

\item{K}{Numeric vector of length equal to the number of samples. The 
elements are integers from 1 to # folds. Indicates which samples belong to 
which fold.}

\item{gwas}{A SnpMatrix object with the GWAS information.}

\item{covars}{A data frame with the covariates. It must contain a column 
'sample' containing the sample IDs, and an additional columns for each 
covariate.}
}
\description{
Take the k-solutions for a combination of hyperparameters, and 
assign a score to it (the larger, the better).
}
\keyword{internal}
