% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scones.R
\name{score_fold}
\alias{score_fold}
\title{Score the solutions of a k-fold}
\usage{
score_fold(gwas, covars, net, selected, criterion, max_prop_snp)
}
\arguments{
\item{gwas}{A SnpMatrix object with the GWAS information.}

\item{covars}{A data frame with the covariates. It must contain a column 
'sample' containing the sample IDs, and an additional columns for each 
covariate.}

\item{net}{An igraph network that connects the SNPs.}

\item{criterion}{String with the function to measure the quality of a split.}

\item{max_prop_snp}{Maximum proportion of SNPs accepted in the model (between 0 and 1). Larger solutions will be discarded.}
}
\description{
Take the k-solutions for a combination of hyperparameters, and 
assign a score to it (the larger, the better).
}
\keyword{internal}
