% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shake_test.R
\name{shake_test}
\alias{shake_test}
\title{Calculate an empirical p-value for shake results.}
\usage{
shake_test(map, net, N = 1e+05)
}
\arguments{
\item{map}{Results from \code{shake}.}

\item{net}{The same SNP network provided to \code{shake}.}

\item{N}{Integer with the name of permutations.}
}
\value{
An empirical p-value for each of the SNP clusters. Please, note that the minimum possible p-value from an empirical distribution is set to 1/(N+1). Clusters composed by a single SNP will not have a empirical p-value.
}
\description{
Do a permutation-based test to assess the statistical significance of each of the clusters obtained through shake. For a cluster of size k, k interconnected SNPs are picked N times, and their joint association score is calculated to come up with an estimation of the distribution.
}
