% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nongwas.R
\name{sigmod.cv_}
\alias{sigmod.cv_}
\title{Find connected explanatory features}
\usage{
sigmod.cv_(X, y, featnames, net)
}
\arguments{
\item{X}{n x d design matrix}

\item{y}{Vector of length n with the outcomes}

\item{featnames}{Vector of length d with the feature names}

\item{net}{An igraph network that connects the SNPs.}
}
\description{
Finds the features maximally associated with a phenotype while 
being connected in an underlying network. Select the hyperparameters by
cross-validation.
}
