% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigmod.R
\name{sigmod}
\alias{sigmod}
\title{Find connected explanatory SNPs}
\usage{
sigmod(gwas, net, eta, lambda, covars = data.frame(), score = "chi2")
}
\arguments{
\item{gwas}{A SnpMatrix object with the GWAS information.}

\item{net}{An igraph network that connects the SNPs.}

\item{eta}{Value of the eta parameter.}

\item{lambda}{Value of the lambda parameter.}

\item{covars}{A data frame with the covariates. It must contain a column 
'sample' containing the sample IDs, and an additional columns for each 
covariate.}

\item{score}{Association score to measure association between genotype and 
phenotype. Possible values: chi2 (default), glm.}
}
\value{
A copy of the \code{SnpMatrix$map} \code{data.frame}, with the 
following additions:
\itemize{
\item{c: contains the univariate association score for every single SNP.}
\item{selected: logical vector indicating if the SNP was selected by SConES 
or not.}
\item{module: integer with the number of the module the SNP belongs to.}
}
}
\description{
Finds the SNPs maximally associated with a phenotype while being
connected in an underlying network.
}
\examples{
gi <- get_GI_network(minigwas, snpMapping = minisnpMapping, ppi = minippi)
sigmod(minigwas, gi, 10, 1)
}
\references{
Liu, Y., Brossard, M., Roqueiro, D., Margaritte-Jeannin, P., 
Sarnowski, C., Bouzigon, E., Demenais, F. (2017). SigMod: an exact and
efficient method to identify a strongly interconnected disease-associated
module in a gene network. Bioinformatics, 33(10), 1536–1544. 
\url{https://doi.org/10.1093/bioinformatics/btx004}
}
