% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nongwas.R
\name{sigmod_}
\alias{sigmod_}
\title{Find connected explanatory features}
\usage{
sigmod_(X, y, featnames, net, eta, lambda)
}
\arguments{
\item{X}{n x d design matrix}

\item{y}{Vector of length n with the outcomes}

\item{featnames}{Vector of length d with the feature names}

\item{net}{An igraph network that connects the SNPs.}

\item{eta}{Value of the eta parameter.}

\item{lambda}{Value of the lambda parameter.}
}
\description{
Finds the features maximally associated with a phenotype while 
being connected in an underlying network.
}
