% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulate_causal_snps}
\alias{simulate_causal_snps}
\title{Simulate causal SNPs}
\usage{
simulate_causal_snps(net, n)
}
\arguments{
\item{net}{An igraph network that connects the SNPs.}

\item{n}{Number of causal SNPs to return.}
}
\value{
A vector with the ids of the simulated causal SNPs.
}
\description{
Selects randomly a clique of interconnected SNPs. If the SNP network contains a "gene" vertex attribute, it tries to pick 
SNPs from the same gene and from, at least, one interactor. Else, it picks SNPs from the largest clique in the network.
}
