% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_causal_snps.R
\name{simulate_causal_snps}
\alias{simulate_causal_snps}
\title{Simulate causal SNPs}
\usage{
simulate_causal_snps(gwas, net, n)
}
\arguments{
\item{net}{An igraph network that connects the SNPs.}

\item{n}{Number of causal SNPs to return.}
}
\value{
A boolean vector with as many elements as SNPs.
}
\description{
Takes the biggest clique of the network, and randomly picks n SNPs from it.
}
