% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scones.R
\name{single_snp_association}
\alias{single_snp_association}
\title{Calculate genotype-phhenotype associations}
\usage{
single_snp_association(
  gwas,
  covars,
  score,
  samples = rep(TRUE, nrow(gwas[["fam"]]))
)
}
\arguments{
\item{gwas}{A SnpMatrix object with the GWAS information.}

\item{covars}{A data frame with the covariates. It must contain a column 
'sample' containing the sample IDs, and an additional columns for each 
covariate.}

\item{score}{Association score to measure association between genotype and 
phenotype. Possible values: chi2 (default), glm.}
}
\value{
A named vector with the association scores.
}
\description{
Calculate the association between genotypes and a phenotype,
adjusting by covariates.
}
\keyword{internal}
