% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{subvert}
\alias{subvert}
\title{Vertices with an attribute}
\usage{
subvert(net, attr, values)
}
\arguments{
\item{net}{An igraph network.}

\item{attr}{An attribute of the vertices.}

\item{values}{Possible values of \code{attr}}
}
\value{
The vertices with attribute equal to any of the values in \code{values}.
}
\description{
Returns the nodes matching some condition.
}
